[RSPEC TESTS] fixed flaky regulation filter test scenarios
