+++ spec/unit/searches/regulations/no_filters_search_spec.rb
@@ -1,0 +1,11 @@
+require "rails_helper"
+
+describe "Regulations search: No filters" do
+
+  include_context "regulations_search_base_context"
+
+  it "should filter" do
+    results = search_results({})
+    expect(results.count).to be_eql(10)
+  end
+end
